requirements.txt
# Generated requirements.txt
# Core dependencies
aiohttp>=3.8.0
# asyncio (built-in)
click>=8.0.0
colorlog>=6.7.0
coverage>=6.0
fastapi>=0.100.0
prometheus-client>=0.16.0
psutil>=5.9.0
psycopg2-binary>=2.9
pydantic>=2.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
python-dotenv>=0.19.0
pyyaml>=6.0
requests>=2.28.0
rich>=13.0.0
sqlalchemy>=2.0
uvicorn>=0.23.0

# Development dependencies (optional)
# black>=22.0.0
# flake8>=5.0.0
# mypy>=1.0.0
